+++ python/iqm/qdmi/__init__.py
@@ -17,12 +17,11 @@
 
 """Python wrapper for exposing the IQM QDMI device library."""
 
-from ._paths import IQM_QDMI_CMAKE_DIR, IQM_QDMI_CONFIG_PATH, IQM_QDMI_INCLUDE_DIR, IQM_QDMI_LIBRARY_PATH
+from ._paths import IQM_QDMI_CMAKE_DIR, IQM_QDMI_INCLUDE_DIR, IQM_QDMI_LIBRARY_PATH
 from ._version import version as __version__
 
 __all__ = [
     "IQM_QDMI_CMAKE_DIR",
-    "IQM_QDMI_CONFIG_PATH",
     "IQM_QDMI_INCLUDE_DIR",
     "IQM_QDMI_LIBRARY_PATH",
     "__version__",